requirements.txt
crewai==0.41.1
langchain>=0.3.1,<0.4.0
langchain-community
langchain-huggingface
gradio==4.8.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy>=1.26.0
